Ajuste de tipo de variável

+++ demo/src/main/java/com/agenda/demo/model/Contatos.java
@@ -33,5 +33,5 @@ public class Contatos implements Serializable {
 	private String endereco;
 
 	@Column(name="TELEFONE")
-	private Integer telefone;
+	private Long telefone;
 }
